Keep custom tmpdir and use existing Lastpass attachments (#570)

+++ keepercommander/importer/lastpass/vault.py
@@ -109,12 +109,15 @@ class Vault(object):
                     os.makedirs(tmpdir)
                 self.tmpdir = os.path.abspath(tmpdir)
 
-        print(f'Downloading {attach_cnt} LastPass attachments:')
+        print(f'Processing {attach_cnt} LastPass attachments:')
         for i, attachment in enumerate(self.attachments):
-            tmp_filename = f'{str(i+1).zfill(attach_cnt_digits)}of{attach_cnt}_{attachment.file_id}'
+            tmp_filename = f'{str(i + 1).zfill(attach_cnt_digits)}of{attach_cnt}_{attachment.file_id}'
             attachment.tmpfile = os.path.join(self.tmpdir, tmp_filename)
-            with fetcher.stream_attachment(session, attachment) as r:
-                with open(attachment.tmpfile, 'wb') as f:
-                    print(f'{i+1}. {attachment.name} ... ', end='', flush=True)
-                    shutil.copyfileobj(r.raw, f)
-                    print('Done')
+            if os.path.isfile(attachment.tmpfile) and os.path.getsize(attachment.tmpfile) == attachment.size:
+                print(f'{i + 1}. Found {attachment.name}')
+            else:
+                with fetcher.stream_attachment(session, attachment) as r:
+                    with open(attachment.tmpfile, 'wb') as f:
+                        print(f'{i + 1}. Downloading {attachment.name} ... ', end='', flush=True)
+                        shutil.copyfileobj(r.raw, f)
+                        print('Done')
